pyproject.toml
[tool.poetry]
name = "snowflake-tools"
version = "0.1.0"
description = ""
authors = ["mtwomey <mtwomey@beakstar.com>"]
readme = "README.md"
packages = [{include = "snowflake_tools", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
snowflake-connector-python = {extras = ["pandas"], version = "^3.7.0"}
toml = "^0.10.2"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
snowflake-mirror-permissions = "snowflake_tools.mirror_permissions:cli"
snowflake-analyze-table = "snowflake_tools.analyze_table:cli"
